+++ apps/harvester/src/Harvester/schedule/index.ts
@@ -4,7 +4,7 @@ import { setSchedule as setScheduleWin32 } from './scheduler-win32';
 import { setSchedule as setScheduleDarwin } from './scheduler-darwin';
 import { log } from '@thecointech/logging';
 
-export function setSchedule(schedule: HarvestSchedule, existing?: HarvestSchedule) {
+export function setSchedule(schedule: HarvestSchedule) {
   // Don't schedule if we're in dev mode
   if (process.env.NODE_ENV === 'development') {
     log.debug("Skipping setSchedule in dev mode");
@@ -13,11 +13,11 @@ export function setSchedule(schedule: HarvestSchedule, existing?: HarvestSchedul
 
   switch (process.platform) {
     case "linux":
-      return setScheduleLinux(schedule, existing);
+      return setScheduleLinux(schedule);
     case "win32":
-      return setScheduleWin32(schedule, existing);
+      return setScheduleWin32(schedule);
     case "darwin":
-      return setScheduleDarwin(schedule, existing);
+      return setScheduleDarwin(schedule);
     default:
       throw new Error("Unsupported platform");
   }